requirements.txt
# Fingerprint Effect Generator

## Dependencies
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.7.0

## Installation
```bash
pip install numpy opencv-python pillow scipy
```

## Quick Test
After installation, test with:
```bash
python examples/basic_usage.py
```
